+++ src/app/customer/CheckoutButtonList.tsx
@@ -3,12 +3,14 @@ import React, { memo, Fragment, FunctionComponent } from 'react';
 
 import { TranslatedString } from '../locale';
 
+import { ApplePayButton } from './customWalletButton';
 import CheckoutButton from './CheckoutButton';
 
 // TODO: The API should tell UI which payment method offers its own checkout button
 export const SUPPORTED_METHODS: string[] = [
     'amazon',
     'amazonpay',
+    'applepay',
     'braintreevisacheckout',
     'chasepay',
     'masterpass',
@@ -19,6 +21,7 @@ export const SUPPORTED_METHODS: string[] = [
     'googlepaycybersourcev2',
     'googlepayorbital',
     'googlepaystripe',
+    'googlepaystripeupe',
 ];
 
 export interface CheckoutButtonListProps {
@@ -64,13 +67,21 @@ const CheckoutButtonList: FunctionComponent<CheckoutButtonListProps> = ({
 
             <div className="checkoutRemote">
                 { supportedMethodIds.map(methodId =>
-                    <CheckoutButton
-                        containerId={ `${methodId}CheckoutButton` }
-                        key={ methodId }
-                        methodId={ methodId }
-                        onError={ onError }
-                        { ...rest }
-                    />
+                    methodId === 'applepay' ?
+                        <ApplePayButton
+                            containerId={ `${methodId}CheckoutButton` }
+                            key={ methodId }
+                            methodId={ methodId }
+                            onError={ onError }
+                            { ...rest }
+                        /> :
+                        <CheckoutButton
+                            containerId={ `${methodId}CheckoutButton` }
+                            key={ methodId }
+                            methodId={ methodId }
+                            onError={ onError }
+                            { ...rest }
+                        />
                 ) }
             </div>
         </Fragment>