+++ src/app/customer/customWalletButton/index.ts
@@ -1,0 +1,1 @@
+export { default as ApplePayButton } from './ApplePayButton';