+++ src/app/payment/mapSubmitOrderErrorMessage.ts
@@ -10,11 +10,17 @@ export default function mapSubmitOrderErrorMessage(
         case 'not_initialized':
             return translate('payment.payment_error');
 
+        case 'custom_provider_execute_error':
+            return translate(error.subtype);
+
         case 'payment_cancelled':
             return translate('payment.payment_cancelled');
 
         case 'payment_method_invalid':
             return translate('payment.payment_method_disabled_error');
+
+        case 'tax_provider_unavailable':
+            return translate('payment.tax_provider_unavailable');
 
         case 'cart_changed':
             return translate('shipping.cart_change_error');