+++ src/app/payment/paymentMethod/AdyenV2CardValidation.spec.tsx
@@ -1,7 +1,7 @@
 import { mount } from 'enzyme';
 import React, { FunctionComponent } from 'react';
 
-import AdyenV2CardValidation, { AdyenV2CardValidationProps } from './AdyenV2CardValidation';
+import AdyenV2CardValidation, { AdyenV2CardValidationProps, AdyenV2CardValidationState } from './AdyenV2CardValidation';
 
 describe('AdyenV2CardValidation', () => {
     let defaultProps: AdyenV2CardValidationProps;
@@ -24,6 +24,7 @@ describe('AdyenV2CardValidation', () => {
 
         const field = container.find('[id="encryptedSecurityCode"]');
 
+        expect(field.hasClass('adyen-checkout__input--error')).toBeFalsy();
         expect(field).toHaveLength(1);
     });
 
@@ -38,5 +39,38 @@ describe('AdyenV2CardValidation', () => {
         const field = container.find('[id="encryptedExpiryDate"]');
 
         expect(field).toHaveLength(1);
+    });
+
+    it('render with empty required fields', () => {
+        defaultProps = {
+            paymentMethodType: 'scheme',
+            shouldShowNumberField: false,
+            verificationFieldsContainerId: 'container',
+            cardValidationState: {} as AdyenV2CardValidationState,
+        };
+
+        const container = mount(<AdyenV2CardValidationTest { ...defaultProps } />);
+
+        const field = container.find('[id="encryptedSecurityCode"]');
+
+        expect(field.hasClass('adyen-checkout__input--error')).toBeTruthy();
+    });
+
+    it('render with invalid fields', () => {
+        defaultProps = {
+            paymentMethodType: 'scheme',
+            shouldShowNumberField: false,
+            verificationFieldsContainerId: 'container',
+            cardValidationState: {
+                fieldType: 'encryptedSecurityCode',
+                valid: false,
+            } as AdyenV2CardValidationState,
+        };
+
+        const container = mount(<AdyenV2CardValidationTest { ...defaultProps } />);
+
+        const field = container.find('[id="encryptedSecurityCode"]');
+
+        expect(field.hasClass('adyen-checkout__input--error')).toBeTruthy();
     });
 });