+++ src/app/payment/paymentMethod/GooglePayPaymentMethod.spec.tsx
@@ -1,6 +1,7 @@
-import { createCheckoutService, CheckoutSelectors, CheckoutService, PaymentMethod } from '@bigcommerce/checkout-sdk';
+import { createCheckoutService, CheckoutSelectors, CheckoutService, PaymentInitializeOptions, PaymentMethod } from '@bigcommerce/checkout-sdk';
 import { mount, ReactWrapper } from 'enzyme';
 import { Formik } from 'formik';
+import each from 'jest-each';
 import { noop } from 'lodash';
 import React, { FunctionComponent } from 'react';
 
@@ -73,7 +74,17 @@ describe('when using Google Pay payment', () => {
             }));
     });
 
-    it('initializes method with required config', () => {
+    each([
+        [PaymentMethodId.AdyenV2GooglePay],
+        [PaymentMethodId.AuthorizeNetGooglePay],
+        [PaymentMethodId.BraintreeGooglePay],
+        [PaymentMethodId.CheckoutcomGooglePay],
+        [PaymentMethodId.CybersourceV2GooglePay],
+        [PaymentMethodId.OrbitalGooglePay],
+        [PaymentMethodId.StripeGooglePay],
+        [PaymentMethodId.StripeUPEGooglePay],
+    ]).it('initializes %s with required config', id => {
+        method.id = id;
         const container = mount(<PaymentMethodTest { ...defaultProps } method={ method } />);
         const component: ReactWrapper<WalletButtonPaymentMethodProps> = container.find(WalletButtonPaymentMethod);
 
@@ -84,12 +95,42 @@ describe('when using Google Pay payment', () => {
 
         expect(checkoutService.initializePayment)
             .toHaveBeenCalledWith(expect.objectContaining({
-                methodId: method.id,
+                methodId: id,
                 gatewayId: method.gateway,
-                [method.id]: {
+                [id]: {
                     walletButton: 'walletButton',
                     onError: defaultProps.onUnhandledError,
+                    onPaymentSelect: expect.any(Function),
                 },
             }));
+    });
+
+    it('reinitializes method once payment option is selected', async () => {
+        const container = mount(<PaymentMethodTest { ...defaultProps } method={ method } />);
+        const component: ReactWrapper<WalletButtonPaymentMethodProps> = container.find(WalletButtonPaymentMethod);
+
+        component.prop('initializePayment')({
+            methodId: method.id,
+            gatewayId: method.gateway,
+        });
+
+        const options: PaymentInitializeOptions = (checkoutService.initializePayment as jest.Mock).mock.calls[0][0];
+
+        const paymentSelectHandler = options.googlepaybraintree?.onPaymentSelect;
+        if (paymentSelectHandler) {
+            paymentSelectHandler();
+        }
+
+        await new Promise(resolve => process.nextTick(resolve));
+
+        expect(checkoutService.deinitializePayment)
+            .toHaveBeenCalledWith({ methodId: method.id });
+        expect(checkoutService.initializePayment)
+            .toHaveBeenCalledWith(expect.objectContaining({
+                methodId: method.id,
+                [method.id]: expect.any(Object),
+            }));
+        expect(checkoutService.initializePayment)
+            .toHaveBeenCalledTimes(2);
     });
 });