+++ src/app/payment/paymentMethod/PaymentMethod.tsx
@@ -26,6 +26,7 @@ import MasterpassPaymentMethod from './MasterpassPaymentMethod';
 import MolliePaymentMethod from './MolliePaymentMethod';
 import MonerisPaymentMethod from './MonerisPaymentMethod';
 import OfflinePaymentMethod from './OfflinePaymentMethod';
+import OpyPaymentMethod from './OpyPaymentMethod';
 import PaymentMethodId from './PaymentMethodId';
 import PaymentMethodProviderType from './PaymentMethodProviderType';
 import PaymentMethodType from './PaymentMethodType';
@@ -36,6 +37,7 @@ import PaypalPaymentsProPaymentMethod from './PaypalPaymentsProPaymentMethod';
 import PPSDKPaymentMethod from './PPSDKPaymentMethod';
 import SquarePaymentMethod from './SquarePaymentMethod';
 import StripePaymentMethod from './StripePaymentMethod';
+import StripeUPEPaymentMethod from './StripeUPEPaymentMethod';
 import VisaCheckoutPaymentMethod from './VisaCheckoutPaymentMethod';
 
 export interface PaymentMethodProps {
@@ -85,6 +87,10 @@ const PaymentMethodComponent: FunctionComponent<PaymentMethodProps & WithCheckou
 
     if (method.gateway === PaymentMethodId.StripeV3) {
         return <StripePaymentMethod { ...props } />;
+    }
+
+    if (method.gateway === PaymentMethodId.StripeUPE) {
+        return <StripeUPEPaymentMethod { ...props } />;
     }
 
     if (method.id === PaymentMethodId.Amazon) {
@@ -151,7 +157,8 @@ const PaymentMethodComponent: FunctionComponent<PaymentMethodProps & WithCheckou
         method.id === PaymentMethodId.CheckoutcomGooglePay ||
         method.id === PaymentMethodId.CybersourceV2GooglePay ||
         method.id === PaymentMethodId.OrbitalGooglePay ||
-        method.id === PaymentMethodId.StripeGooglePay) {
+        method.id === PaymentMethodId.StripeGooglePay ||
+        method.id === PaymentMethodId.StripeUPEGooglePay) {
         return <GooglePayPaymentMethod { ...props } />;
     }
 
@@ -202,7 +209,6 @@ const PaymentMethodComponent: FunctionComponent<PaymentMethodProps & WithCheckou
         method.gateway === PaymentMethodId.Clearpay ||
         method.id === PaymentMethodId.Humm ||
         method.id === PaymentMethodId.Laybuy ||
-        method.id === PaymentMethodId.Opy ||
         method.id === PaymentMethodId.Quadpay ||
         method.id === PaymentMethodId.Sezzle ||
         method.id === PaymentMethodId.Zip ||
@@ -214,6 +220,10 @@ const PaymentMethodComponent: FunctionComponent<PaymentMethodProps & WithCheckou
 
     if (method.type === PaymentMethodProviderType.Offline) {
         return <OfflinePaymentMethod { ...props } />;
+    }
+
+    if (method.id === PaymentMethodId.Opy) {
+        return <OpyPaymentMethod { ...props } />;
     }
 
     if (method.gateway === PaymentMethodId.Mollie) {