+++ src/app/payment/paymentMethod/PaymentMethodId.ts
@@ -12,6 +12,7 @@ enum PaymentMethodId {
     Boleto = 'boleto',
     Bolt = 'bolt',
     Braintree = 'braintree',
+    BraintreeVenmo = 'braintreevenmo',
     AuthorizeNetGooglePay = 'googlepayauthorizenet',
     BraintreeGooglePay = 'googlepaybraintree',
     BraintreeVisaCheckout = 'braintreevisacheckout',
@@ -47,7 +48,9 @@ enum PaymentMethodId {
     Sezzle = 'sezzle',
     SquareV2 = 'squarev2',
     StripeGooglePay = 'googlepaystripe',
+    StripeUPEGooglePay = 'googlepaystripeupe',
     StripeV3 = 'stripev3',
+    StripeUPE = 'stripeupe',
     Zip = 'zip',
 }