+++ src/app/payment/paymentMethod/PaymentMethodTitle.spec.tsx
@@ -24,7 +24,7 @@ describe('PaymentMethodTitle', () => {
     let localeContext: LocaleContextType;
 
     const LOGO_PATHS: { [key: string]: string } = {
-        'paypal-credit': '/img/payment-providers/paypal-credit.png',
+        'paypal-credit': '/img/payment-providers/paypal_commerce_pay_later.svg',
         'visa-checkout': '/img/payment-providers/visa-checkout.png',
         afterpay: '/img/payment-providers/afterpay-badge-blackonmint.png',
         amazon: '/img/payment-providers/amazon-header.png',
@@ -271,6 +271,30 @@ describe('PaymentMethodTitle', () => {
 
         expect(component.find(CreditCardIconList).prop('selectedCardType'))
             .toEqual('mastercard');
+    });
+
+    it('renders a different logo for braintreeVenmo methodId for Paypal', () => {
+        const imageExtension = '.svg';
+        const imageFolder = '/modules/checkout/braintreevenmo/images/';
+        const method = PaymentMethodType.Paypal;
+        const id = PaymentMethodId.BraintreeVenmo;
+        const logoUrl = `${config.cdnPath}${imageFolder}braintree_venmo${imageExtension}`;
+
+        const component = mount(<PaymentMethodTitleTest
+            { ...defaultProps }
+            method={ {
+                ...defaultProps.method,
+                id,
+                method,
+                logoUrl,
+            } }
+        />);
+
+        const expectedPath = `${config.cdnPath}${imageFolder}braintree_venmo${imageExtension}`;
+
+        expect(component.find('[data-test="payment-method-logo"]').prop('src'))
+            .toEqual(expectedPath);
+
     });
 
     it('renders selected credit card type using card number if not using hosted fields', () => {